Paying attention to tests
Tests were succeeding individually but failing when run as a set. It
looks like am seeing messages crossing test boundaries, with
unpredictable results. Have rationalised use of ConnectionMultiplexer --
now passed in to WorkflowManagement,, although that exposes some
redis-y-ness to the client; and explicitly unsubscribe to channels in a
naive Dispose.