#9 first step -- can now fetch workflow details
Generally useful; you don't need to wait for end of workflow to call it.
However, all that's being returned is information that's available at
client, because we can't generate timestamps in Lua in Redis. Needs a
little thought -- could this just be stored at client? Believe the
answer to that is that this is part of the transactional exchange at
point of workflow push, and task pop/complete.