Strategies for recovery from failure
Tighten this up a bit. Default behaviour now is to automatically restart
any task that a component marked as being run by a component with the
same identifier. Left up to the client to ensure that only one client
with each identifier is running at any time, or to instantiate without
Behaviours.AutoRestart.